converts cluster replication attributes from fragment to relationship
