+++ ui/app/serializers/cluster.js
@@ -1,6 +1,9 @@
 import RESTSerializer, { EmbeddedRecordsMixin } from '@ember-data/serializer/rest';
 import { assign } from '@ember/polyfills';
 import { decamelize } from '@ember/string';
+import IdentityManager from '../utils/identity-manager';
+
+const uuids = new IdentityManager();
 
 export default RESTSerializer.extend(EmbeddedRecordsMixin, {
   keyForAttribute: function (attr) {
@@ -9,6 +12,22 @@ export default RESTSerializer.extend(EmbeddedRecordsMixin, {
 
   attrs: {
     nodes: { embedded: 'always' },
+    dr: { embedded: 'always' },
+    performance: { embedded: 'always' },
+  },
+
+  setReplicationId(data) {
+    if (data) {
+      data.id = data.cluster_id || uuids.fetch();
+    }
+  },
+
+  normalize(modelClass, data) {
+    // embedded records need a unique value to be stored
+    // set id for dr and performance to cluster_id or random unique id
+    this.setReplicationId(data.dr);
+    this.setReplicationId(data.performance);
+    return this._super(modelClass, data);
   },
 
   pushPayload(store, payload) {
